requirements.txt
aws-cdk-lib>=2.0.0
jsii>=1.0.0
constructs>=10.0.0
